Solution to Input Two Integers and Print Their Sum

+++ Input Two Integers and Print Their Sum@@6788d5012db25efec946f93d/code.c
@@ -7,5 +7,5 @@ char* welcome() {
 int main() {
    int a,b;
    scanf("%d%d",a,b);
-   printf("Sum: ",(int)(a+b));
+   printf("Sum: ",a+b);
 }
